+++ fabric-convention-tags-v2-bridge/src/main/java/dev/loaderbridge/fabric/api/tag/convention/mixin/TagKeyMixin.java
@@ -1,0 +1,8 @@
+package dev.loaderbridge.fabric.api.tag.convention.mixin;
+
+import net.fabricmc.fabric.api.tag.FabricTagKey;
+import net.minecraft.tags.TagKey;
+import org.spongepowered.asm.mixin.Mixin;
+
+@Mixin(TagKey.class)
+public interface TagKeyMixin extends FabricTagKey { }